Extract InventoryIncrementalJobAPI and extend InventoryIncrementalJobPublicAPI for common usage (#21152)
* Extract more methods to InventoryIncrementalJobPublicAPI; Add InventoryIncrementalJobAPI

* Add PipelineDataConsistencyChecker

* Rename AbstractInventoryIncrementalJobAPIImpl
